2008-04-15  Felipe Erias Morandeira  <femorandeira@maemo.org>
	* src/main.cc (installed_package_activated) : If the user taps on a
	system update, a banner "ai_ni_unable_to_uninstall_system_update" is
	shown. This is for consistency with the changes on 2008-04-14.



git-svn-id: https://garage.maemo.org/svn/hildon-app-mgr/trunk@1113 e55152ac-3a23-0410-980c-f88739b87167

+++ src/main.cc
@@ -1577,7 +1577,10 @@ installed_package_selected (package_info *pi)
 static void
 installed_package_activated (package_info *pi)
 {
-  uninstall_package_flow (pi);
+  if (pi->flags & pkgflag_system_update)
+    irritate_user (_("ai_ni_unable_to_uninstall_system_update"));
+  else
+    uninstall_package_flow (pi);
 }
 
 GtkWidget *
